+++ elementapi/services/ElementApiService.php
@@ -23,4 +23,16 @@ class ElementApiService extends BaseApplicationComponent
 	{
 		$this->raiseEvent('onBeforeSendData', $event);
 	}
+
+	/**
+	 * Fires an 'onAfterSendData' event.
+	 *
+	 * @param Event $event
+	 *
+	 * @return null
+	 */
+	public function onAfterSendData(Event $event)
+	{
+		$this->raiseEvent('onAfterSendData', $event);
+	}
 }