+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Element API
 
+## Unreleased
+
+### Added
+- Added the `elementApi.onAfterSendData` event.
+
 ## 1.6.0 - 2017-05-25
 
 ### Added